requirements.txt
python-dotenv>=1.0.1
requests>=2.32.3
beautifulsoup4>=4.12.3
lxml>=5.3.0
tqdm>=4.66.5
pyyaml>=6.0
tenacity>=9.0.0
python-dateutil>=2.9.0
langdetect>=1.0.9
matplotlib>=3.8
seaborn>=0.13
datasketch>=1.6

polars>=1.12.0
pandas>=2.2.3
datasets>=3.1.0

torch>=2.5.0
transformers>=4.46.0
sentence-transformers>=3.3.0

llama-index>=0.12.0
qdrant-client>=1.12.0
tiktoken>=0.7

wandb>=0.18.7
jupyter>=1.1.1
ipykernel>=6.29.5

google-genai>=1.0
groq>=0.11
anthropic>=0.40

ruff>=0.7.4
pytest>=8.3.3
